requirements.txt
# Core AI and Workflow Dependencies
langchain-openai>=0.1.0
pydantic>=2.6.0
openai>=1.0.0

# Database
asyncpg>=0.29.0
python-dotenv>=1.0.0

# Utilities
typing-extensions>=4.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# Logging and Debugging
rich>=10.0.0  # Enhanced logging and CLI output

# Development Tools
black>=24.1.0  # Code formatting
isort>=5.13.0  # Import sorting
mypy>=1.8.0   # Static type checking
